Make the authentication flow as a nested graph. This allows isolating the authentication flow more easily into its own task stack and closing that task once it's completed.
