+++ app/src/main/kotlin/nl/eduid/di/assist/AuthenticationAssistant.kt
@@ -121,6 +121,13 @@ class AuthenticationAssistant {
                     VersionedBrowserMatcher.FIREFOX_BROWSER
                 )
             )
+//            builder.setBrowserMatcher(
+//                BrowserAllowList(
+//                    VersionedBrowserMatcher.CHROME_CUSTOM_TAB,
+//                    VersionedBrowserMatcher.SAMSUNG_CUSTOM_TAB,
+//                    VersionedBrowserMatcher.FIREFOX_CUSTOM_TAB
+//                )
+//            )
             builder.setConnectionBuilder(DefaultConnectionBuilder.INSTANCE)
 
             return AuthorizationService(context, builder.build())
