+++ app/src/main/kotlin/nl/eduid/messaging/EduIdMessagingService.kt
@@ -74,7 +74,9 @@ class EduIdMessagingService : FirebaseMessagingService() {
             }
 
             val intent = Intent(Intent.ACTION_VIEW, Uri.parse(challenge)).also { intent ->
-                intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_MULTIPLE_TASK or Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS
+                intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_NEW_DOCUMENT or Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS
+                //When using only FLAG_ACTIVITY_MULTIPLE_TASK a new separate task does not seem to be created for
+//                intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_MULTIPLE_TASK or Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS
             }
             val flags =
                 PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
